refactor: organized ir views and added method to ir batch viewer
